ref(sandbox): split snapshot modules
