+++ packages/junior/tests/unit/sandbox/snapshot/profile.test.ts
@@ -1,0 +1,76 @@
+import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
+
+const { dependenciesMock, postinstallMock } = vi.hoisted(() => ({
+  dependenciesMock: vi.fn(),
+  postinstallMock: vi.fn(),
+}));
+
+vi.mock("@/chat/plugins/catalog-runtime", () => ({
+  pluginCatalogRuntime: {
+    getRuntimeDependencies: dependenciesMock,
+    getRuntimePostinstall: postinstallMock,
+  },
+}));
+
+vi.mock("@/chat/sandbox/runtime-dependencies", () => ({
+  GLOBAL_RUNTIME_DEPENDENCIES: [],
+}));
+
+import { create, isStale } from "@/chat/sandbox/snapshot/profile";
+
+describe("snapshot dependency profile", () => {
+  beforeEach(() => {
+    dependenciesMock.mockReset();
+    postinstallMock.mockReset();
+    dependenciesMock.mockReturnValue([]);
+    postinstallMock.mockReturnValue([]);
+    delete process.env.SANDBOX_SNAPSHOT_REBUILD_EPOCH;
+    delete process.env.SANDBOX_SNAPSHOT_FLOATING_MAX_AGE_MS;
+    vi.useFakeTimers();
+    vi.setSystemTime(new Date("2026-03-10T00:00:00.000Z"));
+  });
+
+  afterEach(() => {
+    vi.useRealTimers();
+  });
+
+  it("expires floating npm selectors but keeps exact versions reusable", () => {
+    dependenciesMock.mockReturnValue([
+      { type: "npm", package: "floating", version: "latest" },
+    ]);
+    const floating = create("node22");
+    expect(floating).not.toBeNull();
+    expect(
+      isStale(floating!, new Date("2026-03-01T00:00:00.000Z").getTime()),
+    ).toBe(true);
+
+    dependenciesMock.mockReturnValue([
+      { type: "npm", package: "exact", version: "1.2.3" },
+    ]);
+    const exact = create("node22");
+    expect(exact).not.toBeNull();
+    expect(
+      isStale(exact!, new Date("2025-03-01T00:00:00.000Z").getTime()),
+    ).toBe(false);
+  });
+
+  it("treats postinstall profiles as floating", () => {
+    postinstallMock.mockReturnValue([
+      { cmd: "agent-browser", args: ["install"] },
+    ]);
+    const profile = create("node22");
+    expect(profile?.floating).toBe(true);
+  });
+
+  it("changes the hash when the rebuild epoch changes", () => {
+    dependenciesMock.mockReturnValue([
+      { type: "npm", package: "example", version: "1.2.3" },
+    ]);
+    process.env.SANDBOX_SNAPSHOT_REBUILD_EPOCH = "epoch-a";
+    const first = create("node22");
+    process.env.SANDBOX_SNAPSHOT_REBUILD_EPOCH = "epoch-b";
+    const second = create("node22");
+
+    expect(first?.hash).not.toBe(second?.hash);
+  });
+});
